backend/requirement.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pypdf==4.0.1
python-docx==1.1.0
pydantic==2.10.0
aiofiles==23.2.1
httpx==0.26.0
reportlab==4.1.0

# OCR dependencies for scanned PDF support
pytesseract==0.3.10
pdf2image==1.17.0
Pillow>=10.0.0